Distinguish between binary and source checksum in the manager.
We have a source_checksum that determines the file paths
for the manager compilation artifact. The source checksum
includes not just source (like SQL and Rust code) but also
build environment like the feldera commit sha we're building
with.

Since source checksum is used as the folder name of the pipeline
binary, every new commit of feldera will make new folders for
project of the same SQL code. This is a problem when we want to
use sccache in the project now (in CI, but ideally everywhere
since it seems to help a lot with compile times) because sccache
uses path names as key for it's cache lookup. Hence a new commit
would always result in 100% cache misses because all the paths
change (e.g., for the precompile step).

I change this here so we have a source_checksum that's really
just sources (UDF and SQL code), and a binary checksum
that adds things about the environment. The source checksum
is used for the rust project paths, and the binary checksum
(which was formerly source_checksum is used to determine
the binary names and for recompile decision by the manager).

Signed-off-by: Gerd Zellweger <mail@gerdzellweger.com>