Enable sccache for build-rust jobs.
This speeds up CI job by a few minutes. Rust build on x86 is down
to 3 min and on arm down to 8 min with this.

I also figured out how to configure how to add sccache to our
docker build precompile step and the integration test but currently
the cache hit rate is 0% because our build paths we form in the
manager are not deterministic so I'm hoping the many-crates-pr
will fix that and we should see some significant speedup then
when we add it.

Signed-off-by: Gerd Zellweger <mail@gerdzellweger.com>